Made studydocs a module
